Refactor track filtering functions
